add static_casts to fix linux build

+++ src/trace_processor/importers/etw/etw_parser.cc
@@ -110,60 +110,66 @@ void EtwParser::ParseMemInfo(int64_t timestamp, ConstBytes blob) {
 
   TrackId zero_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Zero"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.zero_page_count(),
-                                       zero_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.zero_page_count()),
+      zero_page_count_track_id);
 
   TrackId free_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Free"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.free_page_count(),
-                                       free_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.free_page_count()),
+      free_page_count_track_id);
 
   TrackId modified_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Modified"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.modified_page_count(),
-                                       modified_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.modified_page_count()),
+      modified_page_count_track_id);
 
   TrackId modified_no_write_page_count_track_id =
       context_->track_tracker->InternTrack(
           kEtwMeminfoBlueprint, tracks::Dimensions("ModifiedNoWrite"));
-  context_->event_tracker->PushCounter(timestamp,
-                                       meminfo.modified_no_write_page_count(),
-                                       modified_no_write_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.modified_no_write_page_count()),
+      modified_no_write_page_count_track_id);
 
   TrackId bad_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Bad"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.bad_page_count(),
-                                       bad_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.bad_page_count()),
+      bad_page_count_track_id);
 
   TrackId modified_page_count_page_file_track_id =
       context_->track_tracker->InternTrack(
           kEtwMeminfoBlueprint, tracks::Dimensions("ModifiedPageFile"));
-  context_->event_tracker->PushCounter(timestamp,
-                                       meminfo.modified_page_count_page_file(),
-                                       modified_page_count_page_file_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.modified_page_count_page_file()),
+      modified_page_count_page_file_track_id);
 
   TrackId paged_pool_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("PagedPool"));
-  context_->event_tracker->PushCounter(timestamp,
-                                       meminfo.paged_pool_page_count(),
-                                       paged_pool_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.paged_pool_page_count()),
+      paged_pool_page_count_track_id);
 
   TrackId non_paged_pool_page_count_track_id =
       context_->track_tracker->InternTrack(kEtwMeminfoBlueprint,
                                            tracks::Dimensions("NonPagedPool"));
-  context_->event_tracker->PushCounter(timestamp,
-                                       meminfo.non_paged_pool_page_count(),
-                                       non_paged_pool_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.non_paged_pool_page_count()),
+      non_paged_pool_page_count_track_id);
 
   TrackId mdl_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Mdl"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.mdl_page_count(),
-                                       mdl_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.mdl_page_count()),
+      mdl_page_count_track_id);
 
   TrackId commit_page_count_track_id = context_->track_tracker->InternTrack(
       kEtwMeminfoBlueprint, tracks::Dimensions("Commit"));
-  context_->event_tracker->PushCounter(timestamp, meminfo.commit_page_count(),
-                                       commit_page_count_track_id);
+  context_->event_tracker->PushCounter(
+      timestamp, static_cast<double>(meminfo.commit_page_count()),
+      commit_page_count_track_id);
 
   auto standby_page_count_iterator = meminfo.standby_page_counts();
   for (int i = 0; standby_page_count_iterator;
@@ -172,7 +178,8 @@ void EtwParser::ParseMemInfo(int64_t timestamp, ConstBytes blob) {
     TrackId standby_page_count_track_id = context_->track_tracker->InternTrack(
         kEtwMeminfoBlueprint, tracks::Dimensions(base::StringView(name)));
     context_->event_tracker->PushCounter(
-        timestamp, *standby_page_count_iterator, standby_page_count_track_id);
+        timestamp, static_cast<double>(*standby_page_count_iterator),
+        standby_page_count_track_id);
   }
 
   auto repurposed_page_count_iterator = meminfo.repurposed_page_counts();
@@ -182,9 +189,9 @@ void EtwParser::ParseMemInfo(int64_t timestamp, ConstBytes blob) {
     TrackId repurposed_page_count_track_id =
         context_->track_tracker->InternTrack(
             kEtwMeminfoBlueprint, tracks::Dimensions(base::StringView(name)));
-    context_->event_tracker->PushCounter(timestamp,
-                                         *repurposed_page_count_iterator,
-                                         repurposed_page_count_track_id);
+    context_->event_tracker->PushCounter(
+        timestamp, static_cast<double>(*repurposed_page_count_iterator),
+        repurposed_page_count_track_id);
   }
 }
